Extract cli parsing into Cmd.h, Cmd.cpp
