+++ app/embed.cpp
@@ -4,54 +4,18 @@
 
 #include "EmbedAudio.h"
 //#include "File.h"
-#include "Image.h"
-#include "Audio.h"
+//#include "Image.h"
+//#include "Audio.h" 
+#include "Cmd.h"
 
 using namespace std;
 namespace fs = std::filesystem;
 
-static bool bestQuality = true;
-void showUsage(std::string programName) { 
-  fmt::fprintf(cerr, 
-      "Usage: %s [audio_file] [image_file]\n %s %s %s", programName, 
-      "Options:\n", 
-      "\t-h, --help\t\tShow this help message\n", 
-      "\t-f, --fast\t\tLower image quality\n");
-} 
-
-bool meetsReq(int argc, char** argv) {
-  if (argc <= 1 || !(argc > 1 && argc < 3)) {
-    showUsage(argv[0]);
-    throw std::exception();
-  } 
-  return true;
-}
-
-map<int, string> parseOptions(int argc, char** argv) {
-  if (!meetsReq(argc, argv)) { throw std::exception(); }
-
-  fs::path imageFilePath;
-  fs::path audioFilePath;
-
-  for (int i = 0; i < argc; i++) {
-    string arg = string(argv[i]);
-
-    if (arg.compare("-h") || arg.compare("--help")) { showUsage(argv[0]); } 
-    if (arg.compare("-f") || arg.compare("--fast")) { bestQuality = true; }
-    if (Image::isImage(arg)) { imageFilePath = argv[i]; }
-    if (Audio::isAudio(arg)) { audioFilePath = argv[i]; }
-  } 
-
-  if (imageFilePath.empty() || audioFilePath.empty()) { throw std::exception(); }
-
-  map<int, string> result = {{0, imageFilePath}, {1, audioFilePath}};
-  return result; 
-}
-
+static bool bestQuality;
 int main(int argc, char** argv) {
   map<int, string> mediaFiles;
   try {
-    mediaFiles = parseOptions(argc, argv);
+    mediaFiles = parseOptions(argc, argv, bestQuality);
   } catch (const std::exception&) {
     return -1;
   }
